Remove Design stuff
